+++ apina-core/src/test/kotlin/fi/evident/apina/spring/SpringModelReaderTest.kt
@@ -5,6 +5,7 @@ import fi.evident.apina.java.reader.loadClassesFromInheritanceTree
 import fi.evident.apina.model.ApiDefinition
 import fi.evident.apina.model.HTTPMethod
 import fi.evident.apina.model.settings.TranslationSettings
+import fi.evident.apina.model.type.ApiType
 import org.junit.Test
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
@@ -41,6 +42,25 @@ class SpringModelReaderTest {
         val bar = assertNotNull(endpointGroup.endpoints.find { it.name == "bar" }, "bar")
         assertEquals(HTTPMethod.POST, foo.method)
         assertEquals(HTTPMethod.GET, bar.method)
+    }
+
+    @Test
+    @Suppress("unused")
+    fun referenceToGenericClass() {
+        abstract class Bar<A>
+        abstract class Foo<B> : Bar<B>()
+
+        @RestController
+        class ControllerWithVariables : Foo<String>() {
+
+            @GetMapping("/foo")
+            fun foo(): Foo<String> = error("")
+        }
+
+        val model = readModel<ControllerWithVariables>()
+
+        val endpoint = model.endpointGroups.single().endpoints.single()
+        assertEquals(ApiType.Class("Foo"), assertNotNull(endpoint.responseBody))
     }
 
     private inline fun <reified T : Any> readModel(): ApiDefinition {